requirements.txt
altair==5.5.0
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.55.3
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
importlib-resources==6.4.5
Jinja2==3.1.3
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.18.4
networkx==3.2.1
numpy==1.26.3
packaging==24.2
pandas==2.2.3
pillow==10.2.0
protobuf==5.29.1
pyarrow==18.1.0
pydeck==0.9.1
pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
smmap==5.0.1
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
toml==0.10.2
torch==2.5.0+cu118
torchaudio==2.5.0+cu118
torchvision==0.20.0+cu118
tornado==6.4.2
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
zipp==3.21.0
